#2475 Command palette with better actions ux flow for query editor (#2517)
