+++ Src/WitsmlExplorer.Frontend/styles/Icons.tsx
@@ -2,7 +2,6 @@ import { Icon } from "@equinor/eds-core-react";
 import {
   accessible,
   account_circle as accountCircle,
-  filter_alt_active as activeFilter,
   add,
   arrow_down as arrowDown,
   arrow_drop_right as arrowDropRight,
@@ -32,29 +31,33 @@ import {
   favorite_filled as favoriteFilled,
   favorite_outlined as favoriteOutlined,
   filter_alt as filter,
+  filter_alt_active as activeFilter,
   folder_open as folderOpen,
   format_line_spacing as formatLine,
   go_to as goTo,
   in_progress as inProgress,
   info_circle as infoCircle,
-  trending_up as isActive,
   keyboard,
   launch,
   more_vertical as moreVertical,
   new_alert as newAlert,
   paste,
   person,
+  play,
   refresh,
   save,
   search,
   settings,
+  style,
   sync,
   text_field as textField,
+  trending_up as isActive,
   tune,
   update,
   upload,
   world
 } from "@equinor/eds-icons";
+
 const icons = {
   accessible,
   accountCircle,
@@ -100,10 +103,12 @@ const icons = {
   newAlert,
   paste,
   person,
+  play,
   refresh,
   save,
   search,
   settings,
+  style,
   sync,
   textField,
   tune,
